+++ src/models/encoder_classifier.h
@@ -116,6 +116,7 @@ public:
 
     modelFeatures_.insert("transformer-heads");
     modelFeatures_.insert("transformer-no-projection");
+    modelFeatures_.insert("transformer-rnn-projection");
     modelFeatures_.insert("transformer-dim-ffn");
     modelFeatures_.insert("transformer-ffn-depth");
     modelFeatures_.insert("transformer-ffn-activation");
@@ -151,6 +152,12 @@ public:
 
   void push_back(Ptr<EncoderBase> encoder) { encoders_.push_back(encoder); }
   void push_back(Ptr<ClassifierBase> classifier) { classifiers_.push_back(classifier); }
+
+  void load(Ptr<ExpressionGraph> graph,
+            const std::vector<io::Item>& items,
+            bool markedReloaded) override {
+    graph->load(items, markedReloaded && !opt<bool>("ignore-model-config", false));
+  }
 
   void load(Ptr<ExpressionGraph> graph,
             const std::string& name,